Implement fetching a single module
Recently, needed to use neb to "rescue" test content from a server that
was going to be wiped. Unfortunately, could not get individual modules,
only collections. This allows specifying a moduleid, which will then be
retrieved. It does _not_ enable publishing a single module outside the
context of a collection.